refactor: 패키지 생성 및 구분
